Fix: Init should leave the Timer in a trigger-able state
This problem occurred when using Google's colab tool.

Depending on the implementation of the python interpreter, the Thread of the Timer is just created and has not run until the trigger is possible. This may happen if the interpreter does not switch the threads at the end of __init__ of Timer, or if the creating thread is very fast.