Fix: Calling cancel from inside the callback should not produce a deadlock.
It should be possible to cancel the timer cycle from the callback of the timer. Before returning, the cancel function tries to acquire and release the lock to ensure that the timer can be triggered again immediately after cancel returns. Using a normal lock produces a deadlock since the lock is already acquired when the callback is invoked and the lock cannot be released.

Using an RLock prevents the deadlock situation.